+++ src/main/java/com/RaqamiUniverse/RaqamiOnlineShop/request/CreateProductRequest.java
@@ -13,7 +13,7 @@ public class CreateProductRequest {
     private String description;
     private BigDecimal price;
     private String brand;
-    private int inventory;
+    private Long inventory;
     private Double discount;
     private Category category;
 }
